backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
TextBlob==0.17.1
python-dotenv==1.0.0
requests==2.31.0
